Combined the app and verified-users vuex modules. Don't block auth on verified users API call
