+++ src/store/modules/app/mutations.js
@@ -6,6 +6,20 @@ const logMutation = (mutationName, ...args) => {
 }
 
 export default {
+  SET_VERIFIED_USERS_ADDRESS_NAME_MAP(currentState, users) {
+    logMutation('SET_VERIFIED_USERS_ADDRESS_NAME_MAP', users)
+
+    const newAddressNameMap = {}
+
+    users.forEach((user) => {
+      if (!user.name || !user.address) return
+      newAddressNameMap[user.address.toLowerCase()] = user.name
+    })
+
+    currentState.verifiedUsersAddressNameMap = newAddressNameMap
+
+  },
+
   SET_CONFIRM_EMAIL_ADDRESS(currentState, emailAddress) {
     logMutation('SET_CONFIRM_EMAIL_ADDRESS', emailAddress)
