+++ src/store/modules/app/state.js
@@ -1,5 +1,6 @@
 export default () => {
   return {
     confirmEmailAddress: null,
+    verifiedUsersAddressNameMap: null,
   }
 }
