--- src/store/modules/verified-users/state.js
@@ -1,5 +1,0 @@
-export default () => {
-  return {
-    addressNameMap: null,
-  }
-}
